+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -1,4 +1,4 @@
-// Generated using Sourcery 0.16.2 — https://github.com/krzysztofzablocki/Sourcery
+// Generated using Sourcery 0.16.1 — https://github.com/krzysztofzablocki/Sourcery
 // DO NOT EDIT
 
 public let masterRuleList = RuleList(rules: [
@@ -33,6 +33,7 @@ public let masterRuleList = RuleList(rules: [
     DuplicateEnumCasesRule.self,
     DuplicateImportsRule.self,
     DynamicInlineRule.self,
+    EmptyCollectionLiteralRule.self,
     EmptyCountRule.self,
     EmptyEnumArgumentsRule.self,
     EmptyParametersRule.self,