+++ CHANGELOG.md
@@ -14,7 +14,9 @@
 
 #### Bug Fixes
 
-* None.
+* Release memory created for sourcekitd requests.  
+  [Colton Schlosser](https://github.com/cltnschlosser)
+  [#2812](https://github.com/realm/SwiftLint/issues/2812)
 
 ## 0.34.0: Anti-Static Wool Dryer Balls