+++ CHANGELOG.md
@@ -18,7 +18,7 @@
   [Colton Schlosser](https://github.com/cltnschlosser)
   [#2812](https://github.com/realm/SwiftLint/issues/2812)
 
-* Fix `swiftlinit rules` output table formatting.  
+* Fix `swiftlint rules` output table formatting.  
   [JP Simard](https://github.com/jpsim)
   [#2787](https://github.com/realm/SwiftLint/issues/2787)