+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 #### Enhancements
 
-* Type name rules considers SwiftUI template code.
+* Type name rules considers SwiftUI template code.  
   [atfelix](https://github.com/atfelix)
   [#2791](https://github.com/realm/SwiftLint/issues/2791)