+++ CHANGELOG.md
@@ -14,6 +14,11 @@
   [atfelix](https://github.com/atfelix)
   [#2791](https://github.com/realm/SwiftLint/issues/2791)
 
+* Add `contains_over_filter_count` opt-in rule to warn against using
+  expressions like `filter(where:).count > 0` instead of `contains(where:)`.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#2803](https://github.com/realm/SwiftLint/issues/2803)
+
 #### Bug Fixes
 
 * Fixed false positive in `colon` rule inside guard and ternary operator.