+++ CHANGELOG.md
@@ -23,7 +23,8 @@
   expressions like `filter(where:).isEmpty` instead of `contains(where:)`.  
   [Marcelo Fabri](https://github.com/marcelofabri)
 
-* Add `empty_collection_literal` opt-in rule to prefer using `isEmpty` to comparison to `[]` or `[:]`.  
+* Add `empty_collection_literal` opt-in rule to prefer using `isEmpty` to
+  comparison to `[]` or `[:]`.  
   [Colton Schlosser](https://github.com/cltnschlosser)
   [#2807](https://github.com/realm/SwiftLint/issues/2807)