+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Master
 
+This is the last release to support building with Swift 4.2.x.
+
 #### Breaking
 
 * None.